ethdev: fix SFT conntrack destroy
Currently, SFT tracks both TCP and UDP connections. However, when
trying to start a SFT flow with CT enabled, only the TCP connection
tracking allocates additional resources that must be released on the
CT destruction.

When destroying a SFT flow, not only the enabled bit but also the
protocol of the entry should be checked. Or else, it will lead to a
crash when trying to access some uninitialized pointer.

Fixes: 04e58c468f17 ("ethdev: expand SFT library functionality")

Signed-off-by: Bing Zhao <bingz@nvidia.com>