net/mlx5: update running non-root missing capability
When running as non-root several capabilities are required to be set
using "setcap".

Creating flows with group > 0 requires writing to the ICM (internal
context memory).
Due to this, the "cap_sys_rawio" capability is required to be set using
the "setcap".

Added information message when Eswitch manager is not enabled which may
be due to the missing capability

Added documentation for the "cap_sys_rawio" requirement.

Signed-off-by: Shiri Kuzin <shirik@nvidia.com>