net/mlx5: fix unnecessary usleep in conntrack
After posting a WQE to the ASO SQ, it needs some time for the
hardware to finish. In the current implementation, a usleep() is
used inside the loop of checking the hardware status.

10 micro seconds waiting is too long and may introduce extra delay
of the status updating in the hardware, especially for the datapath.

The usleep() function suspends execution of the calling thread
for (at least) usec microseconds. The sleep may be lengthened
slightly by any system activity or by the time spent processing the
call or by the granularity of system timers.

usleep() is also be marked as obsolete in the new POSIX
specification.

By removing the usleep(), the datapath will no longer wait for a
long time before next loop to check the status. This will help to
speed up the communication with hardware.

Fixes: 325a69008904 ("net/mlx5: add ASO CT query implementation")

Signed-off-by: Bing Zhao <bingz@nvidia.com>