debian: ensure buildtype flag is set to release
The auto configure approach used in debs creation used meson macros and
introduces the following meson/dpdk flag: --buildtype=plain.

Documentation defines this flag as setting debug mode to false and
optimization at 0 (on a scale of 0 to 3). The default value in dpdk's
meson.build uses buildtype 'release' which is also in non debug mode but
sets an optimization of 3.

Running a simple regex performance test on Ubuntu OS on ARM with dpdk
built via the debs shows a drop in throughput of around 20% when using
the unoptimized 'plain' build rather than 'release'.

Ensure release mode is used in debs compilation by adding the flag to the
rules file. This will override any 'plain' flag added in auto config.

Signed-off-by: John Hurley <jhurley@nvidia.com>