net/mlx5: update null flow operations error
The mlx5_flow_driver_ops is set with operations according to the
driver type.

The driver type is set in flow_get_drv_type to type MLX5_FLOW_TYPE_MAX,
that has null operations in case that the transfer attribute is set and
Eswitch enable is set.

In this case the null validate operation will be used which will return
an empty error message.

Update the error message to inform of missing SW steering capabilities
in case the null operations are used.

Signed-off-by: Shiri Kuzin <shirik@nvidia.com>