net/mlx5: allow action CT with AGE and COUNT together
Currently PMD rejects flow with action CT and AGE together
because they both use ASO action which leads to use two
ASO actions what is not supported yet.

However after commit [1], if flow with action CT and AGE and COUNT
together it should be allowed because AGE will fall-back to
use HW counter, not ASO hit object, so only CT uses the ASO object.

This patch allows the user to create flow with action CT and AGE and
COUNT together, when:

1. Action COUNT is non shared (and non indirect);
2. Action AGE is non shared (and non indirect);
Or:
3. flow hit ASO is disabled;
Or:
4. It's root table (Verbs engine)

[1]:fbe23c42544b ("net/mlx5: use aging by counter when counter exists")

Signed-off-by: Xiaoyu Min <jackmin@nvidia.com>