net/mlx5: support more than 4M conntrack objects
In order to support more than 4M conntrack objects, the apportionment
of the CT index is changed in this commit.

The maximal CT number supported is changed from 4M to 32M, by using
another 3 more bits. In the meanwhile, the maximal ports supported
is reduced to 32 by removing 3 bits. The 2 MSB is still used as the
indirect action type.

The CT pool management is also extended by using u32 instead of u16
as the type of pool index. More than 64K pools can be supported and
there is still 64 CT objects in one pool due to the bulk number.

Some overflow mechanism is used for the last object in the last pool,
and all the least 25 bits with 0 is used as the last one.

Signed-off-by: Bing Zhao <bingz@nvidia.com>