net/mlx5: use stack for translating action handle
When inserting a flow rule, if there is a handle of indirect action
in the action list, it should be translated back to the direct flow
action in place. Some extra memory space is needed in order to save
the result if there is a translation from action handle. By using
some memory allocation mechanism dynamically, it may hurt the
performance in multi-threading cases, since there will be some lock
usually, just like that in rte_malloc.

The stack space or TLS could be used. By pre-reserving some space,
it will save the time for allocation and free. Right now stack is
chosen due to the fact there will be not too many actions in one
single flow. If the space is not enough, some memory is allocated
and freed, the same as current implementation.

This should also be done for a flow validation without creation,
because the indirect actions translation is changed.

Signed-off-by: Bing Zhao <bingz@nvidia.com>