net/mlx5: fix crash when querying a conntrack
In the current implementation, the conntrack update was asynchronous
and query was synchronous. When querying an CT object, a pointer to
the stack memory was used for copying the data from the registered
buffer to the stack. After the data was copied, the spinlock of the
ASO query could be released immediately for other threads' using, to
reduce the latency of analyzing the data and ensure the data
integrity.

By using the stack, the overhead of malloc / free were saved but
a risk was introduced. When there was a failure of the query request
or some misuse of the CT index, the stack pointer related to the WQE
query request may be invalid after the query returned with an error.
If the stack area is reused by other function, there would be a data
corruption or even a crash.

By changing the input parameters of query, the copy will only be done
when the pointer is valid and not "held in the queue" anymore. This
will get rid of the unnecessary copy.

Right now, only one pending query request is supported for one CT
object.

Fixes: 325a69008904 ("net/mlx5: add ASO CT query implementation")

Signed-off-by: Bing Zhao <bingz@nvidia.com>