net/mlx5: fix metadata calculation in legacy mode
Currently, we store metadata in BE and need to change
it to CPU endian when attaching to mbuf.

For dv_xmeta_en = 2, no bit shift.
For dv_xmeta_en = 1, 16 bits shift.

For the non-vector mode, endian change should be first,
then bit shift.

For the vector mode, the behavior is right. In this update,
we use __builtin_clz instead of __builtin_popcount to
avoid confusing.

Fixes: 23cb163a ("net/mlx5: convert meta register to big-endian")

Signed-off-by: rongwei liu <rongweil@nvidia.com>