net/mlx5: fix the CT reference count check when closing
When closing a device, all the conntrack resources should be cleaned
up forcibly. If some CT is still referenced, a counter is used to
record the number of CTs to provide more information. The _sub_fetch
was wrongly used, and if a CT was freed, the reference count would
get an overflow and be considered as referenced.

The _fetch_sub shall be used instead. If the count is 0, nothing will
be done. Only when the count is equal to or larger than 1, the
freeing process will be executed.

In the meanwhile, the log level should be "debug" because all the CT
resources will try to be freed.

Fixes: 99ea8631374b ("net/mlx5: close CT management structure")

Signed-off-by: Bing Zhao <bingz@nvidia.com>