app/test-regex: fix mismatch when multiple iterations
Fix case where we receive mismatches in case we run the
test for multiple iterations (--nb_iter > 1).

Since we use the same regex ops for enqueue and dequeue,
after the first dequeue there becomes a mismatch between
the user_id field (represents the job id) and the mbuf
field (that holds the job data).

Fixes: de06137cb295 ("app/regex: add RegEx test application")

Signed-off-by: Ady Agbarih <adypodoman@gmail.com>