net/mlx5: fix misuse of errno in conntrack create
When creating a connection track object, the index will be returned
for the following usage. There are some cases that the creation will
get a failure and index 0 is returned, e.g., no available memory for
a new pool. The rte_flow_error is set with the proper errno value
and message. But the negative value shouldn't be returned directly.
Or else, it would be treated as some positive unsigned value and
misused in the following code. The negative errno value would be
translated to a large index and cause invalid memory access.

When the creation gets a failure, the 0 should still be used as the
return value to indicate that no available conntrack object is
created successfully. Then a NULL will be returned to the user and
the information stored in the rte_flow_error could be used to get
the detailed error reason.

Fixes: ab3197cb7918 ("net/mlx5: add actions creating for CT")

Signed-off-by: Bing Zhao <bingz@nvidia.com>