regex/mlx5: fix lake pointer initialization in QP setup
The memory region caching mechanism uses generation number to check
memory event. This number is in the global cache, and every queue points
to it.

In the initialization of the QP here the pointer is not initialized as
required, so access to it may cause segmentation fault.

Initialize the pointer to global generation number.

Fixes: 8f847e078e34 ("regex/mlx5: fix memory region unregistration")

Signed-off-by: Michael Baum <michaelba@nvidia.com>