regex/mlx5: fix miss checking in indirect mkey creation
The cache management mechanism does not release MRs when the DPDK
releases them into the system, but turns on a bit that indicates that
resources are waiting to be released.
When a queue searches for cached resources, it first checks the bit and
releases if necessary.

In creating indirect mkey, the look was not tested. As a result, during
the search the queue may find resources that should have been released
before.

Add a shell function that will check the bit and release if necessary.

Fixes: cda883bbb655 ("regex/mlx5: add dynamic memory registration to datapath")

Signed-off-by: Michael Baum <michaelba@nvidia.com>