net/mlx5: fix RSS selection flags settings
The L3 protocol of the RSS type may be different than the one
defined in the flow.
If the RSS type also includes L4 protocol type, the selection
flags for the RX hash will be set with SPORT/DPORT without
setting SRC/DST IP, but this combination is not supported by
the rte API.

When using indirect RSS action, the flow creation fails,
because it does not match any of the pre-created TIRs.

The fix is to prevent setting the hash flags with SPORT/DPORT
without setting SRC/DST IP. The hash flags will remain 0,
meaning non-RSS processing of the received packets.
In case of indirect rss action, it will match the MLX5_RSS_HASH_NONE
pre-created TIR.
In addition, the queue_num is set to 1 when the hash flags are 0,
but it was implemented only when creating a new TIR. Applied the same
to the RSS desc before checking if it matches a cached TIR object.

Fixes: b1d63d829378 ("net/mlx5: support RSS on src or dst fields only")
Fixes: 5a959cbfa68c ("net/mlx5: share Rx hash queue code")
Cc: stable@dpdk.org

Signed-off-by: Lior Margalit <lmargalit@nvidia.com>