net/mlx5: fix crash when using meter hierarchy
If the last meter in hierarchy has queue action in policy, and has
already been used by a flow before the meter hierarchy firstly being
used, it will lead a crash.

The reason is in this case, it skips the sub_policy creation for
hierarchy meters before the last one.

This fixes this by always calling prepare sub_policy to make sure
every meter in hierarchy can have their sub_policy as long as needed.

Fixes: 1ae2a618d6b3 ("net/mlx5: support meter action in meter policy")

Signed-off-by: Shun Hao <shunh@nvidia.com>