mlx5dr: test-mlx5, Add option to pass ib device name
The test application is just that, a test. As such we can do non standard
flows with it. As we don't depend on the DPDK args to drive the test we
need a private way to pass an ib device on which to run the tests.

"-d <dev_name" can be passed as the last arguemnt, all arguments
before are assumed to be DPDK related.

Signed-off-by: Mark Bloch <mbloch@nvidia.com>