test-regex: ensure iters default to 1 in non perf mode
When the test-regex app is run in non performance mode, the data should
be parsed once and the matches returned. Any iterations value input should
be overwritten. It should also mean that, if the user does not explicitly
set any iterations, the data will still be parsed once.

There is a bug in the clause that enforces the above whereby a pointer for
the performance mode setting is checked as opposed to the boolean pointed
to. This, therefore, always evaluates to true.

Fix this by checking the value of the perf mode bool.

Signed-off-by: John Hurley <jhurley@nvidia.com>