vdpa/mlx5: fix no traffic time detection
The driver supports multiple ways for traffic detection in order to
notify the guest when the HW puts packets in the queues.

The `no traffic time` devarg controls on time passed without any traffic
before moving to no traffic mode.
In this mode, the driver stops polling and uses interrupts to detect
traffic.

Previously the user was requested to supply the time in polling cycles
times which is uncalculable when the user works in full poll mode.

Move these devarg units to micro-seconds.

Fixes: ba618270a267 ("vdpa/mlx5: improve interrupt management")
Cc: stable@dpdk.org

Signed-off-by: Matan Azrad <matan@nvidia.com>