net/mlx5: fix the unset error message during flow insertion
There are a variety of causes that result in a failure of inserting
a flow rule. The parameter with type "struct rte_flow_error *" can
be used by the application to detect the detailed reason of a
failure.

For each error branch, the error structure should have the
information that been set via rte_flow_error_set(), or from the
previous called function returning a failure. Only setting the
rte_errno variable is not enough.

In flow_list_create(), there were two branches that returned directly
just with the rte_errno set. With this fix, the error information is
set for the application to get the detailed information.

An unnecessary rte_errno restoring process is also removed.

Fixes: d85c7b5ea59f ("net/mlx5: split hairpin flows")
Fixes: 3f22e3153fb5 ("net/mlx5: check memory allocation in flow creation")
Fixes: 94b6d8843852 ("net/mlx5: reuse flow id as hairpin id")
Cc: orika@nvidia.com
Cc: asafp@nvidia.com
Cc: axuemingl@nvidia.com

Signed-off-by: Bing Zhao <bingz@nvidia.com>