vdpa/mlx5: fix the UAR allocation return value
Once the mlx5_devx_alloc_uar() returns with NULL pointer, it means
that the UAR allocation has a failure.

In the current implementation for this case, the rte_errno was
already set inside mlx5_devx_alloc_uar(). If the caller deassigned
the rte_errno with errno again, in some cases, an incorrect value of
0 would overwrite proper value. Then the following code would have a
misjudgment and continue running. Eventually, unexpected error would
occur instead of capturing the failure immediately.

Removing the line in the caller and using the value already set in
mlx5_devx_alloc_uar() will prevent the misjudgment.

Fixes: 8395927cdfaf ("vdpa/mlx5: prepare HW queues")
Cc: matan@nvidia.com

Signed-off-by: Bing Zhao <bingz@nvidia.com>