net/mlx5: fix workspace queue memory leak in expansion
From the man page below:
  https://linux.die.net/man/3/realloc
If realloc() fails the original block is left untouched; it is not
freed or moved.

When the expansion of RSS queues storage got a failure, the memory
pointer was set to NULL. In this case, the original block couldn't
be tracked by the driver anymore and it would be leaked.

Introdcuing a local variable to save the pointer returned from
realloc() solves this issue. The queue pointer and number will
remain unchanged when a failure with NULL occurs.

Fixes: 8bb81f2649b1 ("net/mlx5: use thread specific flow workspace")
Cc: xuemingl@nvidia.com

Signed-off-by: Bing Zhao <bingz@nvidia.com>