net/mlx5: fix positive error value in flow translate
When translating a rte flow to mlx5 device flow(s), each item and
action will be translated. The driver is also involved. By
definition, when an error occurs, a negative value should be
returned with rte_errno set to the proper errno number.

In case the driver returns opposite value by mistake, the error
codition will be changed to non-zero but not a negative value. Or
else, the flow rule creation process would continue lead to
unpredictable errors.

In the meanwhile, the positive value returned when a MARK action
failure occurred is also fixed.

Fixes: d85c7b5ea59f ("net/mlx5: split hairpin flows")
Fixes: d85c7b5ea59f ("net/mlx5: split hairpin flows")
Fixes: ff9549288e42 ("sft: rte registers selection & fixes")
Cc: getelson@nvidia.com
Cc: orika@nvidia.com

Signed-off-by: Bing Zhao <bingz@nvidia.com>