net/mlx5: create ASO conntrack dummy steering action
In the driver, an unique structure is created to hold the cross
ports steering information for a ASO Devx object. All the 64
conntrack in the bulk will share it.

In some race condition, one thread was trying to destroy an ASO
action while another thread is trying to use another conntrack
action in the same bulk. Once the action to be destroyed was the
last one in this bulk, the structure would be freed and another
thread might access some invalid address like 0.

Adding lock or mutex will hurt the overall performance. By creating
a dummy action for a single Devx object, the shared structure will
not be freed during the runtime and then there is no invalid address
accessing.

Signed-off-by: Bing Zhao <bingz@nvidia.com>