net/mlx5: add flow port configure callback
The commit creates the flow port queue and the job descriptors.

In the new API the queue will be specified to create/destroy the
flows.
The job descriptor size is equal to the queue size, and the job
descriptors are popped from LIFO to convey the flow information
during flow create/destroy. So when polling the result, the flow
information will be extracted from the descriptor and after that
the job descriptor will be push back to the LIFO.

Signed-off-by: Suanming Mou <suanmingm@nvidia.com>