net/mlx5: workaround VXLAN match failure in root table
In a recent update, MISC5 support was added to match the VxLAN
header with the reserved 8 bits.

[1] - "drivers: add VXLAN header the last 8-bits matching support"

It depends on the capability bit exposed by the firmware. Due to the
fact that an adjustment in the driver was missed for MISC5 support,
the rule insertion on the root table would get a failure.

Adding a workaround to ignore the use of MISC5 in the root table
will make the flow rule insertion succeed, but no last 8-bits will
be supported on the root table now.

Signed-off-by: Rongwei Liu <rongweil@nvidia.com>
Signed-off-by: Bing Zhao <bingz@nvidia.com>