mlx5dr: Support dynamic definer selection
The idea is to create a small field copy (fc) array. Pre calculation
consists of converting all the items to header layout format, this
format is used to find the most fit definer.
Once we have the definer layout we keep:
Bit offset, Byte location, Bit mask

which are used to set the field to the definer, similar to MLX5_SET.
The limitation is that single fields cannot be broken into multiple
non continuous bytes, or across different DWs, this is due to the
set function that set a single field based on the DW location.

Performance:

Before the code:
 7181.744740 Insertion
 7196.955040 Deletion

After the code:
 7188.045759 Insertion
 7197.966437 Deletion

Signed-off-by: Alex Vesker <valex@nvidia.com>