net/mlx5: fix mempool allocation for MPRQ
A private data struct is appended after the mempool structure.
For MPRQ the mempool was allocated without receiving the space for the
private data.
An attempt to access the private data to read the flags field resulted
in the application crash.
The fix is to pass the proper private_data_size to the mempool creation
method.

Fixes: 7d6bf6b866b8 ("net/mlx5: add Multi-Packet Rx support")
Cc: stable@dpdk.org

Signed-off-by: Lior Margalit <lmargalit@nvidia.com>