net/mlx5: add item template management callbacks
The template defines flows that have the same matching fields
but with different matching values.
For example, matching on 5 tuple TCP flow, the template will be
(eth(null) + IPv4(source + dest) + TCP(s_port + d_port) while
the values for each rule will be different.

Due to the item template can be used in different domains, the
items will only be cached in item template create stage, while
the template is binded to a dedicated table, the HW criteria
will be created and saved to the table. And different tables
may create the same criteria and will not shared between each
other in order to have better performance.

This commit adds item template management callbacks.

Signed-off-by: Suanming Mou <suanmingm@nvidia.com>