net/mlx5: add action template management callbacks
The action template holds a list of action types that will be
used together on the same rule. The template's actions instances
will be created only when the template bind to the dedicated
group. And the created actions will be saved to each individual
group in order for best performance. The actions will not be
shared with each other.

This commit adds the action template management callbacks which
caches the rte flow action list.

Signed-off-by: Suanming Mou <suanmingm@nvidia.com>