common/mlx5: fix class combination validation
[ upstream commit 288d7c3fa6e8bfca9e1e79a9712ec85d7179b3ce ]

The common probe function gets as a user argument the classes it should
create, and checks whether the combination is valid.

In case the device already exists, it checks the integration of the
above with the classes that the device has.
However, the function does not check the combination when the device
does not exist and it has to create it.

Check if the combination is valid for all cases.

Fixes: f787cb8dc80d ("common/mlx5: add common device driver")
Cc: stable@dpdk.org

Signed-off-by: Michael Baum <michaelba@nvidia.com>
Acked-by: Matan Azrad <matan@nvidia.com>

Signed-off-by: Xiaoyu Min <jackmin@nvidia.com>